Move NPUGuard.h from npu/core to csrc/npu
