+++ csrc/npu/NPUEvent.h
@@ -4,9 +4,9 @@
 #include <utility>
 #include "csrc/core/Macros.h"
 #include "csrc/npu/NPUStream.h"
+#include "csrc/npu/NPUGuard.h"
 #include "npu/acl/include/acl/acl.h"
 #include "npu/core/NPUException.h"
-#include "npu/core/NPUGuard.h"
 #include "npu/core/sys_ctrl/npu_sys_ctrl.h"
 
 namespace c10_npu {
