+++ csrc/npu/NPUGuard.h
