+++ torch_npu/csrc/npu/Stream.cpp
@@ -2,12 +2,14 @@
 #include <structmember.h>
 #include <torch/csrc/Device.h>
 #include <torch/csrc/THP.h>
-#include "npu/core/NPUGuard.h"
-#include "npu/core/sys_ctrl/npu_sys_ctrl.h"
-
-#include "npu/acl/include/acl/acl.h"
+#include "csrc/npu/NPUGuard.h"
 #include "torch_npu/csrc/npu/Module.h"
 #include "torch_npu/csrc/npu/Stream.h"
+
+#include "npu/core/sys_ctrl/npu_sys_ctrl.h"
+#include "npu/acl/include/acl/acl_rt.h"
+#include "npu/acl/include/acl/acl.h"
+#include "npu/acl/include/acl/acl_base.h"
 
 PyObject* THNPStreamClass = nullptr;
